… hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/map/MapKeySetMessageTask.java
… hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/map/MapKeySetWithPredicateMessageTask.java
… hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/map/MapValuesMessageTask.java
… hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/map/MapValuesWithPagingPredicateMessageTask.java
… hazelcast/src/main/java/com/hazelcast/client/impl/protocol/task/map/MapValuesWithPredicateMessageTask.java
… hazelcast/src/main/java/com/hazelcast/client/impl/protocol/template/MapCodecTemplate.java
… hazelcast/src/main/java/com/hazelcast/client/impl/protocol/template/ResponseTemplate.java
… hazelcast/src/main/java/com/hazelcast/map/impl/MapPortableHook.java
… hazelcast/src/main/java/com/hazelcast/map/impl/client/AbstractMapQueryRequest.java
… hazelcast/src/main/java/com/hazelcast/map/impl/client/MapEntrySetRequest.java
… hazelcast/src/main/java/com/hazelcast/map/impl/client/MapKeySetRequest.java
… hazelcast/src/main/java/com/hazelcast/map/impl/client/MapValuesRequest.java
… hazelcast/src/test/java/com/hazelcast/map/MapClientRequestTest.java