JMH test results .

```
New client

Benchmark                 Mode  Cnt    Score    Error  Units
PTest.entrySet           thrpt    5  481.196 ± 30.155  ops/s
PTest.entrySetPredicate  thrpt    5  481.303 ± 36.378  ops/s
PTest.keyset             thrpt    5  582.044 ± 26.960  ops/s
PTest.keysetPredicate    thrpt    5  593.332 ± 25.451  ops/s
PTest.values             thrpt    5  578.870 ± 31.411  ops/s
PTest.valuesPredicate    thrpt    5  589.302 ± 11.811  ops/s

Old Client

Benchmark                 Mode  Cnt    Score    Error  Units
PTest.entrySet           thrpt    5  471.154 ± 16.446  ops/s
PTest.entrySetPredicate  thrpt    5  488.504 ± 11.056  ops/s
PTest.keyset             thrpt    5  536.753 ± 29.754  ops/s
PTest.keysetPredicate    thrpt    5  524.883 ± 35.712  ops/s
PTest.values             thrpt    5  549.128 ± 19.177  ops/s
PTest.valuesPredicate    thrpt    5  529.379 ± 21.721  ops/s
```