@Donnerbart 2015-10-06
I don't see the migration of `ClientMapProxy.values()` to `ClientMapProxy.values(TruePredicate.INSTANCE)` in this PR. Do you plan to do this in a next step?

I'm asking because I took care about adpating the JavaDoc of `IMap` and the reference manual about the scope of the `QueryResultSizeExceededException`.