@sancar 2015-10-06
@Donnerbart I solved it diffrently since this one more sutiable to client. I do not chage the message send over the wire. It is more like ValuesRequest extends QueryRequest in the old client perspective. Same code runs on the server side.