@Donnerbart 2015-10-06
So it will use the queries internally and may throw the exception if too much data is returned? Actually we'll see, since there are nightly tests checking those methods, and they will fail if you changed it under the hood :D