@sancar 2015-10-08
@pveentjer @Donnerbart Can we merge this pr now ? I have added the performance measurements with jmh at the top.