@Donnerbart 2015-10-08
I was just worried about the queries not being used (since this would be a difference to the old client and the possible exception). Since this seems to be resolved I have no objections :)

After this PR being merged I expect the `ClientMapUnboundReturnValues_BasicTest` nightly test to fail (PR builder doesn't execute it). I will be happy to adapt the test, since I know what to do there.