@pveentjer 2015-10-09
Looks good. Do we have enough tests for clients?

E.g. map[(1,peter), (2.peter)]

when you ask values.. how many peters do you get.

Also the paging predicate needs enough coverage.

In 3.7 we should clean up the client tests so they can reuse the server tests for logical testing. No need to repeat.