@sancar 2015-10-09
@pveentjer I ported the changes from MapQueryRequest. I did not think on that that much. And there are tests in ClientMapBasicTest with and without predicates and checking the sizes.
And I see tests about PagingPredicate in ClientMapIssueTest.